pyproject.toml
[project]
name = "advent-of-code"
version = "0.2.0"
description = "solve advent of code puzzle"
authors = [
    {name = "cdubos", email = "contact@cdubos.fr"},
]
dependencies = [
    "httpx>=0.25,<0.26",
    "typer>=0.9.0",
    "python-decouple>=3.8",
    "beautifulsoup4>=4.12.2",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["flit_core>=3.2,<4"]
build-backend = "flit_core.buildapi"

[project.optional-dependencies]
2023 = [
    "numpy>=1.26.2",
]

[project.scripts]
advent-of-code = "advent_of_code.app:app"

[tool.pdm.dev-dependencies]
dev = [
    "pre-commit>=3.5.0",
    "ipython>=8.18.1",
]
